+++ src/lib/forms/index.ts
@@ -1,5 +1,6 @@
 import Form, { createForm } from './form.svelte';
 import FieldController from './fieldController.svelte';
 import FormError from './formError.svelte';
+import Input from './input.svelte';
 
-export { Form, createForm, FieldController, FormError };
+export { Form, createForm, FieldController, FormError, Input };