+++ src/routes/[[lang]]/(auth)/login/reset-password/+page.svelte
@@ -2,7 +2,7 @@
 	import Illustration from '$lib/components/molecules/Illustration.svelte';
 	import { Form, createForm } from '$lib/forms';
 	import { m } from '$lib/i18n';
-	import Input from '$lib/ionic/forms/input.svelte';
+	import Input from '$lib/forms/input.svelte';
 	import { arrowForward } from 'ionicons/icons';
 	import { z } from 'zod';
 	import background from '$lib/assets/bg-4.svg';